Several improvements:
- factorization of all the utils
- cleanup of the workflow
- no validation is done, only default are provided, but they can be overridden
- export findmuon data
- distortions
