+++ examples/data/LiF.cif
@@ -1,0 +1,33 @@
+data_image0
+_chemical_formula_structural       LiFLiFLiFLiF
+_chemical_formula_sum              "Li4 F4"
+_cell_length_a       4.08343
+_cell_length_b       4.08343
+_cell_length_c       4.08343
+_cell_angle_alpha    90
+_cell_angle_beta     90
+_cell_angle_gamma    90
+
+_space_group_name_H-M_alt    "P 1"
+_space_group_IT_number       1
+
+loop_
+  _space_group_symop_operation_xyz
+  'x, y, z'
+
+loop_
+  _atom_site_type_symbol
+  _atom_site_label
+  _atom_site_symmetry_multiplicity
+  _atom_site_fract_x
+  _atom_site_fract_y
+  _atom_site_fract_z
+  _atom_site_occupancy
+  Li  Li1       1.0  0.00000  0.00000  0.00000  1.0000
+  F   F1        1.0  0.00000  0.50000  0.00000  1.0000
+  Li  Li2       1.0  0.00000  0.50000  0.50000  1.0000
+  F   F2        1.0  0.00000  0.00000  0.50000  1.0000
+  Li  Li3       1.0  0.50000  0.00000  0.50000  1.0000
+  F   F3        1.0  0.50000  0.50000  0.50000  1.0000
+  Li  Li4       1.0  0.50000  0.50000  0.00000  1.0000
+  F   F4        1.0  0.50000  0.00000  0.00000  1.0000
