+++ examples/data/LiF_allsites.cif
@@ -1,0 +1,38 @@
+# generated using pymatgen
+data_LiHF
+_symmetry_space_group_name_H-M   'P 1'
+_cell_length_a   4.08342800
+_cell_length_b   4.08342800
+_cell_length_c   4.08342800
+_cell_angle_alpha   90.00000000
+_cell_angle_beta   90.00000000
+_cell_angle_gamma   90.00000000
+_symmetry_Int_Tables_number   1
+_chemical_formula_structural   LiHF
+_chemical_formula_sum   'Li4 H4 F4'
+_cell_volume   68.08864750
+_cell_formula_units_Z   4
+loop_
+ _symmetry_equiv_pos_site_id
+ _symmetry_equiv_pos_as_xyz
+  1  'x, y, z'
+loop_
+ _atom_site_type_symbol
+ _atom_site_label
+ _atom_site_symmetry_multiplicity
+ _atom_site_fract_x
+ _atom_site_fract_y
+ _atom_site_fract_z
+ _atom_site_occupancy
+  Li  Li0  1  0.00000000  0.00000000  0.00000000  1.0
+  F  F1  1  0.00000000  0.50000000  0.00000000  1.0
+  Li  Li2  1  0.00000000  0.50000000  0.50000000  1.0
+  F  F3  1  0.00000000  0.00000000  0.50000000  1.0
+  Li  Li4  1  0.50000000  0.00000000  0.50000000  1.0
+  F  F5  1  0.50000000  0.50000000  0.50000000  1.0
+  Li  Li6  1  0.50000000  0.50000000  0.00000000  1.0
+  F  F7  1  0.50000000  0.00000000  0.00000000  1.0
+  H  H8  1  0.20100000  0.00100000  0.20100000  1
+  H  H9  1  0.20100000  0.20100000  0.20100000  1
+  H  H10  1  0.20100000  0.20100000  0.40100000  1
+  H  H11  1  0.40100000  0.20100000  0.40100000  1
