+++ pyproject.toml
@@ -5,10 +5,10 @@ build-backend = "flit_core.buildapi"
 [project]
 name = "aiida-muon"
 dynamic = ["version"] # read from aiida_muon/__init__.py
-description = "aiida-muon is allows to find candiate muon implantation sites and hyperfine field by DFT supercell relaxations and from further symmetry and kinetics analysis.  "
+description = "aiida-muon is allows to find candidate muon implantation sites and hyperfine field by DFT supercell relaxations and from further symmetry and kinetics analysis.  "
 authors = [{name = "Muon group Parma"},{name = "Laboratory for Materials Simulations PSI"}]
 readme = "README.md"
-license = {file = "LICENSE"}
+license = {file = "LICENSE.txt"}
 classifiers = [
     "Programming Language :: Python",
     "Intended Audience :: Science/Research",
